requirements.txt
python-telegram-bot~=21.6
python-dotenv~=1.0.1
openai~=1.51.2
dependency-injector~=4.42.0
requests~=2.32.3
pymongo~=4.10.1